6.1.5 added smart component - создал абстрактный класс smart - вынес в него логику обновления и навешивания обработчиков
